feat: migrate dekan dashboard routes and views from AuthController to DekanController
